refactored BehaviorBuilder

+++ src/gameEngine/actors/behaviors/ClosestRangeAttack.java
@@ -1,12 +1,10 @@
 package gameEngine.actors.behaviors;
 
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
-import javafx.geometry.Point2D;
 import gameEngine.actors.BaseActor;
 import gameEngine.actors.BaseEnemy;
 import gameEngine.actors.BaseTower;
+import java.util.List;
+import javafx.geometry.Point2D;
 /**
  * Example of BaseAttack, where the actor shoots at the nearest opposing actor
  * 
@@ -15,7 +13,7 @@ import gameEngine.actors.BaseTower;
  */
 public class ClosestRangeAttack extends RangeAttack {
 
-    public ClosestRangeAttack (int attackSpeed) {
+    public ClosestRangeAttack (double attackSpeed) {
         super(attackSpeed);
     }
